# hotel-order-ui/src/app/services/models/page-response-location-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { LocationResponse } from '../models/location-response';
export interface PageResponseLocationResponse {
  content?: Array<LocationResponse>;
  empty?: boolean;
  first?: boolean;
  last?: boolean;
  number?: number;
  numberOfElements?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}