# hotel-order-ui/src/app/services/models/service-request-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { LocationResponse } from '../models/location-response';
export interface ServiceRequestResponse {
  completedDate?: string;
  createdDate?: string;
  id?: string;
  location?: LocationResponse;
  locationId?: string;
  serviceStatus?: 'PENDING' | 'CANCELED' | 'COMPLETED';
  serviceType?: 'CALL' | 'BILL';
}