# hotel-order-ui/src/openapi/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenApi Specification ",
        "description": "OpenApi Documentation for Hotel Order Management",
        "termsOfService": "term of service",
        "contact": {
            "name": "Habte",
            "email": "habtadugna@gmail.com"
        },
        "license": {
            "name": "Licence name ",
            "url": "http://url licence"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8088/api/v1",
            "description": "Local ENV"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/service-requests/update-status/{request_id}": {
            "put": {
                "tags": [
                    "service-requests"
                ],
                "operationId": "updateServiceRequestStatus",
                "parameters": [
                    {
                        "name": "request_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/locations/rooms/update-status/{room-id}": {
            "put": {
                "tags": [
                    "locations"
                ],
                "operationId": "updateRoomStatus",
                "parameters": [
                    {
                        "name": "room-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "OCCUPIED",
                                "ONCHANGE",
                                "READY",
                                "DND",
                                "OOO"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/service-requests": {
            "get": {
                "tags": [
                    "service-requests"
                ],
                "operationId": "getAllServiceRequests",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ServiceRequestResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "service-requests"
                ],
                "operationId": "createServiceRequest",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ServiceRequestRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/locations": {
            "get": {
                "tags": [
                    "locations"
                ],
                "operationId": "getAllOrderLocations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LocationResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "locations"
                ],
                "operationId": "saveOrderLocation",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LocationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IdResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/orders/locations/{location-id}": {
            "get": {
                "tags": [
                    "orders"
                ],
                "operationId": "getOrdersByLocation",
                "parameters": [
                    {
                        "name": "location-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrderResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/locations/{location-id}": {
            "get": {
                "tags": [
                    "locations"
                ],
                "operationId": "getLocationById",
                "parameters": [
                    {
                        "name": "location-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LocationResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "locations"
                ],
                "operationId": "deleteLocationById",
                "parameters": [
                    {
                        "name": "location-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/locations/search/number/{location-number}": {
            "get": {
                "tags": [
                    "locations"
                ],
                "operationId": "searchLocationByNumber",
                "parameters": [
                    {
                        "name": "location-number",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "location-type",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LocationResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/locations/page": {
            "get": {
                "tags": [
                    "locations"
                ],
                "operationId": "getPageOfLocations",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseLocationResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/locations/page/{location-type}": {
            "get": {
                "tags": [
                    "locations"
                ],
                "operationId": "getPageOfLocationByType",
                "parameters": [
                    {
                        "name": "location-type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseLocationResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "IdResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    }
                }
            },
...
            "ServiceRequestRequest": {
                "required": [
                    "locationId",
                    "serviceType"
                ],
                "type": "object",
                "properties": {
                    "locationId": {
                        "type": "string"
                    },
                    "serviceType": {
                        "type": "string"
                    }
                }
            },
            "MessageResponse": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    }
                }
            },
...
            "LocationRequest": {
                "required": [
                    "address",
                    "description",
                    "number",
                    "status",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "type": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    }
                }
            },
...
            "LocationResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "TABLE",
                            "ROOM"
                        ]
                    },
                    "address": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "OCCUPIED",
                            "ONCHANGE",
                            "READY",
                            "DND",
                            "OOO"
                        ]
                    },
                    "description": {
                        "type": "string"
                    },
                    "lastModifiedBy": {
                        "type": "string"
                    },
                    "lastModifiedDate": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "ServiceRequestResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "locationId": {
                        "type": "string"
                    },
                    "location": {
                        "$ref": "#/components/schemas/LocationResponse"
                    },
                    "serviceStatus": {
                        "type": "string",
                        "enum": [
                            "PENDING",
                            "CANCELED",
                            "COMPLETED"
                        ]
                    },
                    "serviceType": {
                        "type": "string",
                        "enum": [
                            "CALL",
                            "BILL"
                        ]
                    },
                    "createdDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "completedDate": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "OrderDetailResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "orderId": {
                        "type": "string"
                    },
                    "itemId": {
                        "type": "string"
                    },
                    "itemName": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "quantity": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "type": {
                        "type": "string"
                    },
                    "note": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    }
                }
            },
            "OrderResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "totalPrice": {
                        "type": "number"
                    },
                    "orderType": {
                        "type": "string",
                        "enum": [
                            "DINE_IN",
                            "TAKEAWAY"
                        ]
                    },
                    "orderDetails": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderDetailResponse"
                        }
                    },
                    "location": {
                        "$ref": "#/components/schemas/LocationResponse"
                    },
                    "locationId": {
                        "type": "string"
                    },
                    "note": {
                        "type": "string"
                    },
                    "createdDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "orderStatus": {
                        "type": "string",
                        "enum": [
                            "PENDING",
                            "VERIFIED",
                            "OnPROCESS",
                            "READY",
                            "SERVED",
                            "COMPLETED",
                            "CANCELED",
                            "BillREADY",
                            "BillREQUEST"
                        ]
                    }
                }
            },
...
            "PageResponseLocationResponse": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LocationResponse"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "numberOfElements": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "last": {
                        "type": "boolean"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "empty": {
                        "type": "boolean"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT auth description",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}